add static_cast

+++ include/flashinfer/attention/mla_hopper.cuh
@@ -826,7 +826,8 @@ __global__ __launch_bounds__(KTraits::NUM_THREADS) void BatchMLAPageAttentionHop
           1 - (kv_start / CTA_TILE_KV);
 
       int mask_tile_idx =
-          (CAUSAL ? min(kv_end, kv_len - q_len + packed_qo_start / num_heads) : kv_end) /
+          (CAUSAL ? min(kv_end, kv_len - q_len + static_cast<uint32_t>(packed_qo_start) / num_heads)
+                  : kv_end) /
               CTA_TILE_KV -
           (kv_start / CTA_TILE_KV);
